Update libraries, address typos and modify access tokens
This commit addresses several changes across different files. It updates the versions of various libraries in `package.json` and `package-lock.json` for improved security and functionality. A typographical error in the `revalidate` API call in `route.ts` is fixed. It also adds a fallback access token to `RoomDetails.tsx`, and modifies the way features are mapped in `UpdateRoom.tsx`. Lastly, it hardcodes a localhost URL with a token in `revalidate.ts` for testing purpose.